underwater-prototype/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
numpy
scikit-image
pillow
aiofiles
jinja2
sqlalchemy
python-dotenv
